+++ src/source_files/file_management/DiskFlusher.cpp
@@ -28,7 +28,7 @@ void DiskFlusher::flush_to_disk_periodically() {
 	}
 }
 
-unsigned int DiskFlusher::append_data_to_end_of_file(const std::string& key, const std::string& path, void* data, unsigned long total_bytes, bool flush_immediatelly) {
+unsigned long long DiskFlusher::append_data_to_end_of_file(const std::string& key, const std::string& path, void* data, unsigned long total_bytes, bool flush_immediatelly) {
 	return this->write_data_to_file(key, path, data, total_bytes, -1, flush_immediatelly);
 }
 
@@ -79,10 +79,10 @@ void DiskFlusher::flush_metadata_updates_to_disk(const std::string& key, const s
 	);
 }
 
-unsigned int DiskFlusher::write_data_to_file(const std::string& key, const std::string& path, void* data, unsigned long total_bytes, long long pos, bool flush_immediatelly) {
+unsigned long long DiskFlusher::write_data_to_file(const std::string& key, const std::string& path, void* data, unsigned long total_bytes, long long pos, bool flush_immediatelly) {
 	std::unique_lock<std::mutex> lock(this->flush_mut);
 
-	unsigned int begin_written_pos = this->fh->write_to_file(
+	unsigned long long begin_written_pos = this->fh->write_to_file(
 		key,
 		path,
 		total_bytes,
